pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embedded-serial-bridge"
version = "0.1.0"
description = "Embedded serial bridge with HDLC framing and simple CLI"
readme = "README.md"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "pyserial>=3.5",
  "click>=8.0",
  "tomli>=2.0; python_version<'3.11'",
]

[project.urls]
Homepage = "https://github.com/your/repo"

[project.scripts]
embedded-serial-bridge = "embedded_serial_bridge.cli:main"

[tool.setuptools]
packages = ["embedded_serial_bridge"]
